fix(#1731): move refresh_node from watch to reload
